requirements.txt
python-telegram-bot==21.6
pytz==2024.2
langgraph==0.2.45
langchain==0.3.7
langchain-openai==0.2.6
openai==1.54.0
neo4j==5.26.0
weaviate-client==3.26.7
duckduckgo-search==6.3.7
lxml==5.3.0
beautifulsoup4==4.12.3
apscheduler==3.10.4
python-dotenv==1.0.1
httpx==0.27.2
tenacity==9.0.0
aiohttp==3.10.10
requests==2.32.3
feedparser==6.0.11
trafilatura==1.12.2
readability-lxml==0.8.1
scrapetube
yt-dlp
youtube-transcript-api
